+++ crates/proof-impl/evm-ee-stf/Cargo.toml
@@ -7,7 +7,7 @@ version = "0.3.0-alpha.1"
 workspace = true
 
 [dependencies]
-alpen-reth-evm.workspace = true
+# alpen-reth-evm.workspace = true
 alpen-reth-primitives.workspace = true
 strata-primitives.workspace = true
 strata-state.workspace = true
