+++ crates/reth/evm/src/lib.rs
@@ -1,11 +1,11 @@
 //! This crate holds commong evm changes shared between native and prover runtimes
 //! and should not include any dependencies that cannot be run in the prover.
-pub mod constants;
-mod utils;
+// pub mod constants;
+// mod utils;
 
-pub use evm::AlpenEvmPrecompiles;
-pub use utils::collect_withdrawal_intents;
+// pub use evm::AlpenEvmPrecompiles;
+// pub use utils::collect_withdrawal_intents;
 
-pub mod api;
+// pub mod api;
 pub mod evm;
-mod precompiles;
+// mod precompiles;
