+++ crates/reth/exex/Cargo.toml
@@ -29,7 +29,6 @@ reth-trie-common = { workspace = true, features = ["eip1186"] }
 revm-primitives.workspace = true
 rsp-mpt.workspace = true
 rsp-primitives.workspace = true
-serde_json.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
