+++ crates/reth/exex/src/alloy2reth.rs
@@ -18,13 +18,12 @@ impl IntoReth<RethWithdrawal> for AlloyWithdrawal {
 }
 
 /// A trait to convert from Alloy genesis to RSP genesis.
-pub trait IntoRspGenesis {
-    fn try_into_rsp(self) -> Result<rsp_primitives::genesis::Genesis, serde_json::Error>;
+pub trait IntoRspChainConfig {
+    fn into_rsp(self) -> rsp_primitives::genesis::Genesis;
 }
 
-impl IntoRspGenesis for alloy_genesis::Genesis {
-    fn try_into_rsp(self) -> Result<rsp_primitives::genesis::Genesis, serde_json::Error> {
-        let genesis_str = serde_json::to_string(&self)?;
-        Ok(rsp_primitives::genesis::Genesis::Custom(genesis_str))
+impl IntoRspChainConfig for alloy_genesis::ChainConfig {
+    fn into_rsp(self) -> rsp_primitives::genesis::Genesis {
+        rsp_primitives::genesis::Genesis::Custom(self)
     }
 }
