+++ crates/reth/node/src/evm.rs
@@ -15,19 +15,31 @@ impl<Node> ExecutorBuilder<Node> for AlpenExecutorBuilder
 where
     Node: FullNodeTypes<Types: NodeTypes<ChainSpec = ChainSpec, Primitives = EthPrimitives>>,
 {
-    type EVM = EthEvmConfig<AlpenEvmFactory>;
+    type EVM = EthEvmConfig;
 
-    type Executor = BasicBlockExecutorProvider<Self::EVM>;
-
-    async fn build_evm(
-        self,
-        ctx: &BuilderContext<Node>,
-    ) -> eyre::Result<(Self::EVM, Self::Executor)> {
-        let evm_config =
-            EthEvmConfig::new_with_evm_factory(ctx.chain_spec(), AlpenEvmFactory::default());
-        Ok((
-            evm_config.clone(),
-            BasicBlockExecutorProvider::new(evm_config),
-        ))
+    async fn build_evm(self, ctx: &BuilderContext<Node>) -> eyre::Result<Self::EVM> {
+        let evm_config = EthEvmConfig::new(ctx.chain_spec());
+        Ok(evm_config.clone())
     }
 }
+
+// impl<Node> ExecutorBuilder<Node> for AlpenExecutorBuilder
+// where
+//     Node: FullNodeTypes<Types: NodeTypes<ChainSpec = ChainSpec, Primitives = EthPrimitives>>,
+// {
+//     type EVM = EthEvmConfig<AlpenEvmFactory>;
+
+//     type Executor = BasicBlockExecutorProvider<Self::EVM>;
+
+//     async fn build_evm(
+//         self,
+//         ctx: &BuilderContext<Node>,
+//     ) -> eyre::Result<(Self::EVM, Self::Executor)> {
+//         let evm_config =
+//             EthEvmConfig::new_with_evm_factory(ctx.chain_spec(), AlpenEvmFactory::default());
+//         Ok((
+//             evm_config.clone(),
+//             BasicBlockExecutorProvider::new(evm_config),
+//         ))
+//     }
+// }
